verify collateral zk
