+++ solana/automata-dcap-framework/programs/automata-on-chain-pccs/src/types/mod.rs
@@ -1,0 +1,1 @@
+pub mod zk;
